[Rename] dir `viewComponent` to `component`
