+++ src/editor/view/component/editorViewEventHandler.ts
